✨♻️ Support security at the operation level
